Move ext to ext folder
