+++ OutOfSchool/OutOfSchool.BusinessLogic/Extensions/StreamExtensions.cs
@@ -1,0 +1,40 @@
+﻿using System.Text;
+
+using Newtonsoft.Json;
+
+namespace OutOfSchool.BusinessLogic.Extensions;
+
+public static class StreamExtensions
+{
+    public static T ReadAndDeserializeFromJson<T>(this Stream stream)
+    {
+        ArgumentNullException.ThrowIfNull(stream);
+
+        if (!stream.CanRead)
+        {
+            throw new NotSupportedException("Can't read this stream");
+        }
+
+        using var streamReader = new StreamReader(stream);
+
+        return JsonConvert.DeserializeObject<T>(streamReader.ReadToEnd());
+    }
+
+    public static void SerializeToJsonAndWrite<T>(this Stream stream, T objectToWrite)
+    {
+        ArgumentNullException.ThrowIfNull(stream);
+
+        if (!stream.CanRead)
+        {
+            throw new NotSupportedException("Can't write to this stream");
+        }
+
+        using var streamWriter = new StreamWriter(stream, new UTF8Encoding(), CommunicationConstants.BufferSize, true);
+        using var jsonTextWriter = new JsonTextWriter(streamWriter);
+
+        var jsonSerializer = new JsonSerializer();
+
+        jsonSerializer.Serialize(jsonTextWriter, objectToWrite);
+        jsonTextWriter.Flush();
+    }
+}
